health: Refactor and simplify, drop healthd shims
This change drastically simplifies the code, which had been written
under the assumption that AOSP would drop healthd soon. The wrapper
libraries make little sense, just drop them.
