+++ hardware/health/Android.bp
@@ -6,9 +6,13 @@ cc_binary {
     cflags: [
         "-Wall",
         "-Werror",
+        "-Wno-unused",
+        "-fexceptions",
     ],
     srcs: [
         "HealthService.cpp",
+        "CycleCountBackupRestore.cpp",
+        "LearnedCapacityBackupRestore.cpp",
     ],
     static_libs: [
         "android.hardware.health@2.0-impl",
@@ -18,23 +22,20 @@ cc_binary {
         "libhealthservice",
         // Storage is needed by the HAL definition, use default
         "libhealthstoragedefault",
-        "libhealthboardcommon.sony",
-        "libhealthd.sony",
     ],
     shared_libs: [
+        "android.hardware.health@2.0",
         "libbase",
         "libcutils",
         "libhidlbase",
         "libhidltransport",
         "libhwbinder",
+        "liblog",
         "libutils",
-        "android.hardware.health@2.0",
     ],
-    // HealthService.cpp never uses healthd_board_* functions, so the compiler
-    // drops them. But BatteryMonitor.cpp and Health.cpp depend on those
-    // functions to be defined(not only declared).
-    whole_static_libs: [
-        "libhealthd.sony",
+    header_libs: [
+        // We only need BatteryService.h
+        "libbatteryservice_headers",
     ],
 
     // Only availbale in current Android Q master.
@@ -47,72 +48,4 @@ cc_binary {
     overrides: [
         "healthd",
     ],
-}
-
-cc_library_static {
-    // Vendor-specific functions
-    name: "libhealthboardcommon.sony",
-    // recovery_available is not yet available in Pie, need to use master or wait for Q
-    //recovery_available: true,
-    vendor_available: true,
-    cflags: [
-        "-Wall",
-        "-Werror",
-        "-fexceptions",
-    ],
-    srcs: [
-        "HealthBoardCommon.cpp",
-        "CycleCountBackupRestore.cpp",
-        "LearnedCapacityBackupRestore.cpp",
-    ],
-    shared_libs: [
-        "android.hardware.health@2.0",
-        "libbase",
-        "libhidlbase",
-        "libhidltransport",
-        "libhwbinder",
-        "liblog",
-        "libcutils",
-        "libutils",
-    ],
-    header_libs: [
-        // We only need BatteryService.h for the common lib
-        "libbatteryservice_headers",
-    ],
-    // Needed to find own headers
-    local_include_dirs: ["include"],
-}
-
-cc_library_static {
-    // Shim lib that maps legacy healthd_board_* hooks to vendor lib functions
-    // Needed until the upstream health rework is complete
-    name: "libhealthd.sony",
-    // recovery_available is not yet available in Pie, need to use master or wait for Q
-    //recovery_available: true,
-    vendor_available: true,
-    cflags: [
-        "-Wall",
-        "-Werror",
-    ],
-    shared_libs: [
-        "android.hardware.health@2.0",
-        "libbase",
-        "libhidlbase",
-        "libhidltransport",
-        "libhwbinder",
-        "liblog",
-        "libutils",
-        "libcutils",
-    ],
-    srcs: [
-        "libhealthd_board.cpp",
-    ],
-    header_libs: [
-        // To access healthd/healthd.h and BatteryService.h in libhealthd_board.h
-        "libhealthd_headers",
-    ],
-    // Needed to find own headers
-    local_include_dirs: ["include"],
-    // Let HealthService find headers
-    export_include_dirs: ["include"],
 }
